requirements.txt
torch>=2.3
torchvision>=0.18
transformers>=4.44
accelerate>=0.34
bitsandbytes>=0.43
datasets>=2.20
einops
sentencepiece
safetensors
pillow
tqdm
gradio>=4.44
